minor ui change
