+++ lib/screen/home/view/home_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:phone_comparison_app/widgets/app_bar.dart';
+import '../../../utils/constant/image_asset.dart';
 import '../widget/announcement_list.dart';
-import '../widget/home_banner.dart';
 import '../widget/home_drawer.dart';
 
 class HomeScreen extends StatefulWidget {
@@ -30,6 +30,23 @@ class _HomeScreenState extends State<HomeScreen> {
               child: AnnouncementsList(),
             ),
           ],
+        ),
+      ),
+    );
+  }
+}
+
+class HomeBanner extends StatelessWidget {
+  const HomeBanner({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      height: 200,
+      decoration: BoxDecoration(
+        image: DecorationImage(
+          image: AssetImage(ImageAsset().homeBanner),
+          fit: BoxFit.cover,
         ),
       ),
     );
